Factored out more error types, more reorganization.
